% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineStocks}
\alias{combineStocks}
\title{combineStocks}
\usage{
combineStocks(...)
}
\arguments{
\item{...}{One or more output objects, to be combined to list of models.}
}
\description{
This function takes model objects (class \code{outputs}) of JJM and generate an object
with combined models.
}
\examples{
model <- runJJM(modelName = "mod2.4")
}

