% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagnostics}
\alias{diagnostics}
\title{Generate Assessment plots from single model}
\usage{
diagnostics(outputObject, ...)
}
\arguments{
\item{outputObject}{Object ob class outputs.}

\item{...}{Extra arguments}
}
\description{
Function to generate plots from results of readJJM function
}
\examples{
model <- readJJM(modelName = "mod2.4")
diagnostics(outputObject = model)
}

