% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jjmTools-main.R
\docType{package}
\name{jjmTools-package}
\alias{jjmTools}
\alias{jjmTools-package}
\title{Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.}
\description{
Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.
}
\author{
Ricardo Oliveros-Ramos, Wencheng Lau-Medrano, Giancarlo Morón, Josymar Torrejón-Magallanes and
Niels Hintzen
}
\seealso{
Joint Jack Mackerel Repository
\code{\link{https://github.com/SPRFMO/jack_mackerel}}
}
\keyword{jjmTools}

