% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kobe}
\alias{kobe}
\title{Kobe plot}
\usage{
kobe(model, Blim = 0.5, Flim = 1.5, Bref = 1, Fref = 1, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{model}{Name for new model created with the \code{readJJM} function.}

\item{Blim}{Limit reference point for B/B_MSY, default=0.5.}

\item{Flim}{Limit reference point for F/F_MSY, default=1.5.}

\item{Bref}{Reference point for B/B_MSY, default=1.}

\item{Fref}{Reference point for F/F_MSY, default=1.}

\item{xlim}{'x' axis limits.}

\item{ylim}{'y' axis limits.}

\item{...}{Additional parameters passed to plot.}
}
\description{
This function create a kobe plot from JJM  model outputs
}
\examples{
kobe(model)
}

