% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readJJM}
\alias{readJJM}
\title{Function to read models and list if models and generate results}
\usage{
readJJM(modelName, path = "", ...)
}
\arguments{
\item{modelName}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{...}{}
}
\description{
Function to read models and list if models and generate results
}
\examples{
readJJM(modelName = "mod2.4")
readJJM(modelName = paste0("mod2.", 1:4))
}

