% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jjmTools-main.R
\name{runJJM}
\alias{runJJM}
\title{Run a JJM model}
\usage{
runJJM("models", path = ".", input = "", useGuess = FALSE, exec = "",
  guess = NULL, iprint = 100, wait = TRUE, parallel = FALSE,
  temp = NULL, ...)
}
\arguments{
\item{models}{String with the name of the models to be run.}

\item{path}{Folder where the .ctl is located.}

\item{input}{Folder where the .dat is located.}

\item{exec}{Executable version. jjm2014.exe: One stock, jjm2015.exe: Multi-stock}

\item{useGuess}{boolean, to use an initial guess for the parameters?}

\item{guess}{File with the initial guess for the parameters. If \code{NULL}, will use \code{model.par} in the output folder.}

\item{iprint}{iprint parameter for the JJM model, 100 by default.}

\item{wait}{boolean, wait for the model to finish? Forced to be TRUE.}

\item{temp}{character, path for a temporal directory to run models, if \code{NULL} a temporal folder is automaticaly created.}

\item{...}{Arguments passed from \code{system} function.}
}
\description{
Function to run one or several JJM models
}
\examples{
model = runJJM(model = "model2014ms", path = "config", input = "input", exec= "jjm2015.exe")
}

