% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jjmTools-main.R
\name{combineStocks}
\alias{combineStocks}
\title{combineStocks}
\usage{
combineStocks(..., modelName = NULL)
}
\arguments{
\item{...}{One or more output objects, to be combined to list of models.}

\item{model}{Name for new model. If \code{NULL} (default) a temporal name will be used.}
}
\description{
This function takes model objects (class \code{outputs}) of JJM and generate a model
with SD and mean combined.
}
\examples{
mod1 <- runJJM(modelName = "mod2.1")
mod2 <- runJJM(modelName = "mod2.2")
mod3 <- runJJM(modelName = "mod2.3")

combinedMod_123 <- combineStocks(mod1, mod2, mod3, modelName = NULL)
}

