\name{jjmTools-package}
\alias{jjmTools-package}
\alias{jjmTools}
\docType{package}
\title{
Tools to process and get results from Joint Jack Mackerel outputs.
}
\description{
Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.
}
\details{
\tabular{ll}{
Package: \tab jjmTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-15\cr
License: \tab TBD\cr
}
}
\author{
Ricardo Oliveros-Ramos <roliveros@imarpe.gob.pe>
}
\references{
}
\keyword{ jjmTools }
\seealso{
Joint Jack Mackerel Repository
\code{\link{https://github.com/SPRFMO/jack_mackerel}}
}
\examples{
}
