% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jjmTools-main.R
\docType{package}
\name{jjmTools-package}
\alias{jjmTools}
\alias{jjmTools-package}
\title{Tools to process and get results from Joint Jack Mackerel (JJM) model outputs.}
\description{
Graphics and diagnostics tools for SPRFMO's Joint Jack Mackerel model.
}
\details{
\tabular{ll}{ Package: \tab jjmTools\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2014-08-15\cr License: \tab TBD\cr }
}
\author{
Ricardo Oliveros-Ramos, Wencheng Lau-Medrano, Giancarlo Moron and
Niels Hintzen
}
\seealso{
Joint Jack Mackerel Repository
\code{\link{https://github.com/SPRFMO/jack_mackerel}}
}
\keyword{jjmTools}

