% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jjmTools-main.R
\name{readJJM}
\alias{readJJM}
\title{Read a model or list of models}
\usage{
readJJM(model, path = "", input = "", version = "", ...)
}
\arguments{
\item{model}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{version}{2014: One-stock, 2015: Multi-stock}

\item{...}{Extra arguments}
}
\description{
Function to read models and list if models and generate results
}
\examples{
 readJJM("model2014", path="config", input="input", version="2014")
}

