% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readJJM}
\alias{readJJM}
\title{Read a model or list of models}
\usage{
readJJM(model, path = "", output = "arc", modelName = model, ...)
}
\arguments{
\item{model}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{modelName}{alias for \code{model} (to be deprecated).}

\item{...}{Extra arguments}
}
\description{
Function to read models and list if models and generate results
}
\examples{
readJJM(model = "mod2.4")
}

