% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runJJM}
\alias{runJJM}
\title{Run a JJM model}
\usage{
runJJM(modelName, path = "", wait = TRUE, ...)
}
\arguments{
\item{modelName}{String with the name of model that will be readed or run.}

\item{path}{Directory where the 'admb' folder is located.}

\item{...}{Arguments passed from \code{system} function.}
}
\description{
Function to run a model (o list of models) of JJM and generate inputs files
}
\examples{
model <- runJJM(modelName = "mod2.4")
}

