% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runJJM}
\alias{runJJM}
\title{Function to run a model (o list of models) of JJM and generate inputs files}
\usage{
runJJM(modelName, path = "", wait = TRUE, ...)
}
\description{
Function to run a model (o list of models) of JJM and generate inputs files
}

